hide sidebar pin button when side force collapsed
The button to pin the sidebar is when disabled of no
use, and could even confuse due to the button's state
not reflecting the sidebar's state.

To solve this confusion, the button can be hidden,
when the sidebar is forced into the collapsed mode
on smaller screens or window widths.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>