appdata: Improve appdata for AppStream 1.0
- Add the `<developer><name>` tag
- Mark the `<developer_name>` tag as deprecated
- Mark the developer_name as untranslatable
- Improve appstreamcli parameters
- Activate meson tests on flatpak manifest

### Appdata: Fix my mistake

"Icons and categories

If there’s a type="desktop-id" launchable, they get pulled from it.
Most of the icon not found errors with the flathub builder can be
traced down to the launchable value not matching the desktop file name.

Don’t set them in the AppData unless you want to override them
(even though then it might be a better idea to patch the desktop file
itself)."

More information: https://docs.flathub.org/docs/for-app-authors/appdata-guidelines/#icons-and-categories