add ability to display window controls in every toolbar view
Using Adw.HeaderBar as the top child of an
Adw.ToolbarView ensures that window controls are always
displayed correctly. This is an issue for the small
window sizes (<= 450px width), where the sidebar does not
display them. This is, however, the normal behaviour for
those kinds of sidebars covering the whole app's primary
window.

Additionally, the left most pane visible in medium and
large window mode (<= 750px and 751px+) would need
manually added Gtk.WindowControls for side: start to
supported close buttons on the left for completeness
purposes.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>